+++ 3rdparty/kdiff3/fileaccess.cpp
@@ -1720,7 +1720,7 @@ void ProgressDialog::setInformation(const QString &info, double dCurrent, bool b
 
     ProgressLevelData &pld = m_progressStack.back();
     pld.m_dCurrent = dCurrent;
-    int level = m_progressStack.size();
+    int level = static_cast<int>(m_progressStack.size());
 
     if (level == 1) {
         m_pInformation->setText(info);
@@ -1739,7 +1739,7 @@ void ProgressDialog::setInformation(const QString &info, bool bRedrawUpdate)
     }
 
     //ProgressLevelData& pld = m_progressStack.back();
-    int level = m_progressStack.size();
+    int level = static_cast<int>(m_progressStack.size());
 
     if (level == 1) {
         m_pInformation->setText(info);
@@ -1855,7 +1855,7 @@ void ProgressDialog::recalc(bool bUpdate)
 
     m_progressDelayTimer = startTimer(3000);   /* 3 s delay */
 
-    int level = m_progressStack.size();
+    int level = static_cast<int>(m_progressStack.size());
 
     if ((bUpdate && level == 1) || m_t1.elapsed() > 200) {
         if (m_progressStack.empty()) {
